+++ lib/ModelMixins/CatalogMemberMixin.ts
@@ -1,13 +1,15 @@
-import { computed, runInAction } from "mobx";
+import { action, computed, runInAction } from "mobx";
 import AsyncLoader from "../Core/AsyncLoader";
 import Constructor from "../Core/Constructor";
 import isDefined from "../Core/isDefined";
 import Result from "../Core/Result";
+import hasTraits from "../Models/Definition/hasTraits";
 import Model, { BaseModel } from "../Models/Definition/Model";
 import updateModelFromJson from "../Models/Definition/updateModelFromJson";
 import SelectableDimensions, {
   SelectableDimension
 } from "../Models/SelectableDimensions";
+import CatalogMemberReferenceTraits from "../Traits/TraitsClasses/CatalogMemberReferenceTraits";
 import CatalogMemberTraits from "../Traits/TraitsClasses/CatalogMemberTraits";
 import AccessControlMixin from "./AccessControlMixin";
 import GroupMixin from "./GroupMixin";
@@ -203,10 +205,13 @@ namespace CatalogMemberMixin {
 export default CatalogMemberMixin;
 
 /** Convenience function to get user readable name of a BaseModel */
-export function getName(model: BaseModel | undefined) {
+export const getName = action((model: BaseModel | undefined) => {
   return (
     (CatalogMemberMixin.isMixedInto(model) ? model.name : undefined) ??
+    (hasTraits(model, CatalogMemberReferenceTraits, "name")
+      ? model.name
+      : undefined) ??
     model?.uniqueId ??
     "Unknown model"
   );
-}
+});