+++ lib/ReactViews/Map/Panels/HelpPanel/HelpPanel.jsx
@@ -16,6 +16,8 @@ import {
   HelpAction
 } from "../../../../Core/AnalyticEvents/analyticEvents";
 
+export const HELP_PANEL_ID = "help";
+
 @observer
 class HelpPanel extends React.Component {
   static displayName = "HelpPanel";
@@ -29,15 +31,21 @@ class HelpPanel extends React.Component {
 
   constructor(props) {
     super(props);
+    this.state = {
+      isAnimatingOpen: true
+    };
+  }
+
+  componentDidMount() {
+    // The animation timing is controlled in the CSS so the timeout can be 0 here.
+    setTimeout(() => this.setState({ isAnimatingOpen: false }), 0);
   }
 
   render() {
     const { t } = this.props;
     const helpItems = this.props.terria.configParameters.helpContent;
-    const isVisible =
-      this.props.viewState.showHelpMenu &&
-      this.props.viewState.topElement === "HelpPanel";
     const isExpanded = this.props.viewState.helpPanelExpanded;
+    const isAnimatingOpen = this.state.isAnimatingOpen;
     return (
       <Box
         displayInlineBlock
@@ -52,7 +60,7 @@ class HelpPanel extends React.Component {
             : 110};
           transition: right 0.25s;
           transition-timing-function: cubic-bezier(0.4, 0, 0.2, 1);
-          right: ${isVisible ? (isExpanded ? 490 : 0) : -320}px;
+          right: ${isAnimatingOpen ? -320 : isExpanded ? 490 : 0}px;
         `}
       >
         <Box position="absolute" paddedRatio={3} topRight>
@@ -118,19 +126,17 @@ class HelpPanel extends React.Component {
           </Box>
         </Box>
         <Spacing bottom={10} />
-        <Box centered displayInlineBlock fullWidth>
-          <Box displayInlineBlock fullWidth>
-            {helpItems && (
-              <For each="item" index="i" of={helpItems}>
-                <HelpPanelItem
-                  key={i}
-                  terria={this.props.terria}
-                  viewState={this.props.viewState}
-                  content={item}
-                />
-              </For>
-            )}
-          </Box>
+        <Box centered displayInlineBlock fullWidth styledPadding="0 26px">
+          {helpItems && (
+            <For each="item" index="i" of={helpItems}>
+              <HelpPanelItem
+                key={i}
+                terria={this.props.terria}
+                viewState={this.props.viewState}
+                content={item}
+              />
+            </For>
+          )}
         </Box>
       </Box>
     );