+++ test/ReactViews/ToolButtonSpec.tsx
@@ -4,7 +4,6 @@ import Terria from "../../lib/Models/Terria";
 import ViewState from "../../lib/ReactViewModels/ViewState";
 import Icon from "../../lib/Styled/Icon";
 import MapIconButton from "../../lib/ReactViews/MapIconButton/MapIconButton";
-import { ToolButton } from "../../lib/ReactViews/Tool";
 
 const TestComponent = () => <div>Test hello</div>;
 
@@ -19,19 +18,5 @@ describe("ToolButton", function() {
       catalogSearchProvider: undefined,
       locationSearchProviders: []
     });
-  });
-
-  it("renders a map icon button", function() {
-    rendered = TestRenderer.create(
-      // @ts-ignore
-      <ToolButton
-        toolName="Test tool"
-        viewState={viewState}
-        icon={Icon.GLYPHS.bulb}
-        getToolComponent={() => TestComponent as any}
-        t={(key: any) => key}
-      />
-    );
-    rendered.root.findByType(MapIconButton);
   });
 });